% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{CreateAuthUriResponse}
\alias{CreateAuthUriResponse}
\title{Google Identity Toolkit API Objects 
Help the third party sites to implement federated login.}
\usage{
CreateAuthUriResponse(allProviders = NULL, authUri = NULL,
  captchaRequired = NULL, forExistingProvider = NULL, providerId = NULL,
  registered = NULL, sessionId = NULL)
}
\arguments{
\item{allProviders}{all providers the user has once used to do federated login}

\item{authUri}{The URI used by the IDP to authenticate the user}

\item{captchaRequired}{True if captcha is required}

\item{forExistingProvider}{True if the authUri is for user's existing provider}

\item{providerId}{The provider ID of the auth URI}

\item{registered}{Whether the user is registered if the identifier is an email}

\item{sessionId}{Session ID which should be passed in the following verifyAssertion request}
}
\value{
CreateAuthUriResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:23:37
filename: /Users/mark/dev/R/autoGoogleAPI/googleidentitytoolkitv3.auto/R/identitytoolkit_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
CreateAuthUriResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of creating the IDP authentication URL.
}

