% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{DownloadAccountResponse}
\alias{DownloadAccountResponse}
\title{DownloadAccountResponse Object}
\usage{
DownloadAccountResponse(nextPageToken = NULL, users = NULL)
}
\arguments{
\item{nextPageToken}{The next page token}

\item{users}{The user accounts data}
}
\value{
DownloadAccountResponse object
}
\description{
DownloadAccountResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of downloading accounts in batch.
}
