% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{EmailTemplate}
\alias{EmailTemplate}
\title{EmailTemplate Object}
\usage{
EmailTemplate(body = NULL, format = NULL, from = NULL,
  fromDisplayName = NULL, replyTo = NULL, subject = NULL)
}
\arguments{
\item{body}{Email body}

\item{format}{Email body format}

\item{from}{From address of the email}

\item{fromDisplayName}{From display name}

\item{replyTo}{Reply-to address}

\item{subject}{Subject of the email}
}
\value{
EmailTemplate object
}
\description{
EmailTemplate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for an email template.
}
