% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GetAccountInfoResponse}
\alias{GetAccountInfoResponse}
\title{GetAccountInfoResponse Object}
\usage{
GetAccountInfoResponse(users = NULL)
}
\arguments{
\item{users}{The info of the users}
}
\value{
GetAccountInfoResponse object
}
\description{
GetAccountInfoResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of getting account information.
}
