% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GetOobConfirmationCodeResponse}
\alias{GetOobConfirmationCodeResponse}
\title{GetOobConfirmationCodeResponse Object}
\usage{
GetOobConfirmationCodeResponse(email = NULL, oobCode = NULL)
}
\arguments{
\item{email}{The email address that the email is sent to}

\item{oobCode}{The code to be send to the user}
}
\value{
GetOobConfirmationCodeResponse object
}
\description{
GetOobConfirmationCodeResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of getting a code for user confirmation (reset password, change email etc.).
}
