% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{GetRecaptchaParamResponse}
\alias{GetRecaptchaParamResponse}
\title{GetRecaptchaParamResponse Object}
\usage{
GetRecaptchaParamResponse(recaptchaSiteKey = NULL, recaptchaStoken = NULL)
}
\arguments{
\item{recaptchaSiteKey}{Site key registered at recaptcha}

\item{recaptchaStoken}{The stoken field for the recaptcha widget, used to request captcha challenge}
}
\value{
GetRecaptchaParamResponse object
}
\description{
GetRecaptchaParamResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of getting recaptcha param.
}
