% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\docType{package}
\name{identitytoolkit_googleAuthR}
\alias{identitytoolkit_googleAuthR}
\alias{identitytoolkit_googleAuthR-package}
\title{Google Identity Toolkit API
Help the third party sites to implement federated login.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:59:27
filename: /Users/mark/dev/R/autoGoogleAPI/googleidentitytoolkitv3.auto/R/identitytoolkit_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
}
}
