% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyCreateAuthUriRequest.customParameter}
\alias{IdentitytoolkitRelyingpartyCreateAuthUriRequest.customParameter}
\title{IdentitytoolkitRelyingpartyCreateAuthUriRequest.customParameter Object}
\usage{
IdentitytoolkitRelyingpartyCreateAuthUriRequest.customParameter()
}
\value{
IdentitytoolkitRelyingpartyCreateAuthUriRequest.customParameter object
}
\description{
IdentitytoolkitRelyingpartyCreateAuthUriRequest.customParameter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The query parameter that client can customize by themselves in auth url. The following parameters are reserved for server so that they cannot be customized by clients: client_id, response_type, scope, redirect_uri, state, oauth_token.
}
\seealso{
Other IdentitytoolkitRelyingpartyCreateAuthUriRequest functions: \code{\link{IdentitytoolkitRelyingpartyCreateAuthUriRequest}},
  \code{\link{relyingparty.createAuthUri}}
}
