% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyCreateAuthUriRequest}
\alias{IdentitytoolkitRelyingpartyCreateAuthUriRequest}
\title{IdentitytoolkitRelyingpartyCreateAuthUriRequest Object}
\usage{
IdentitytoolkitRelyingpartyCreateAuthUriRequest(appId = NULL,
  authFlowType = NULL, clientId = NULL, context = NULL,
  continueUri = NULL, hostedDomain = NULL, identifier = NULL,
  oauthConsumerKey = NULL, oauthScope = NULL, openidRealm = NULL,
  otaApp = NULL, providerId = NULL, sessionId = NULL)
}
\arguments{
\item{appId}{The app ID of the mobile app, base64(CERT_SHA1):PACKAGE_NAME for Android, BUNDLE_ID for iOS}

\item{authFlowType}{Explicitly specify the auth flow type}

\item{clientId}{The relying party OAuth client ID}

\item{context}{The opaque value used by the client to maintain context info between the authentication request and the IDP callback}

\item{continueUri}{The URI to which the IDP redirects the user after the federated login flow}

\item{hostedDomain}{The hosted domain to restrict sign-in to accounts at that domain for Google Apps hosted accounts}

\item{identifier}{The email or federated ID of the user}

\item{oauthConsumerKey}{The developer's consumer key for OpenId OAuth Extension}

\item{oauthScope}{Additional oauth scopes, beyond the basid user profile, that the user would be prompted to grant}

\item{openidRealm}{Optional realm for OpenID protocol}

\item{otaApp}{The native app package for OTA installation}

\item{providerId}{The IdP ID}

\item{sessionId}{The session_id passed by client}
}
\value{
IdentitytoolkitRelyingpartyCreateAuthUriRequest object
}
\description{
IdentitytoolkitRelyingpartyCreateAuthUriRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to get the IDP authentication URL.
}
\seealso{
Other IdentitytoolkitRelyingpartyCreateAuthUriRequest functions: \code{\link{relyingparty.createAuthUri}}
}

