% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyDeleteAccountRequest}
\alias{IdentitytoolkitRelyingpartyDeleteAccountRequest}
\title{IdentitytoolkitRelyingpartyDeleteAccountRequest Object}
\usage{

  IdentitytoolkitRelyingpartyDeleteAccountRequest(delegatedProjectNumber = NULL,
  idToken = NULL, localId = NULL)
}
\arguments{
\item{delegatedProjectNumber}{GCP project number of the requesting delegated app}

\item{idToken}{The GITKit token or STS id token of the authenticated user}

\item{localId}{The local ID of the user}
}
\value{
IdentitytoolkitRelyingpartyDeleteAccountRequest object
}
\description{
IdentitytoolkitRelyingpartyDeleteAccountRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to delete account.
}
\seealso{
Other IdentitytoolkitRelyingpartyDeleteAccountRequest functions: \code{\link{relyingparty.deleteAccount}}
}
