% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyDownloadAccountRequest}
\alias{IdentitytoolkitRelyingpartyDownloadAccountRequest}
\title{IdentitytoolkitRelyingpartyDownloadAccountRequest Object}
\usage{

  IdentitytoolkitRelyingpartyDownloadAccountRequest(delegatedProjectNumber = NULL,
  maxResults = NULL, nextPageToken = NULL, targetProjectId = NULL)
}
\arguments{
\item{delegatedProjectNumber}{GCP project number of the requesting delegated app}

\item{maxResults}{The max number of results to return in the response}

\item{nextPageToken}{The token for the next page}

\item{targetProjectId}{Specify which project (field value is actually project id) to operate}
}
\value{
IdentitytoolkitRelyingpartyDownloadAccountRequest object
}
\description{
IdentitytoolkitRelyingpartyDownloadAccountRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to download user account in batch.
}
\seealso{
Other IdentitytoolkitRelyingpartyDownloadAccountRequest functions: \code{\link{relyingparty.downloadAccount}}
}
