% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyGetAccountInfoRequest}
\alias{IdentitytoolkitRelyingpartyGetAccountInfoRequest}
\title{IdentitytoolkitRelyingpartyGetAccountInfoRequest Object}
\usage{

  IdentitytoolkitRelyingpartyGetAccountInfoRequest(delegatedProjectNumber = NULL,
  email = NULL, idToken = NULL, localId = NULL)
}
\arguments{
\item{delegatedProjectNumber}{GCP project number of the requesting delegated app}

\item{email}{The list of emails of the users to inquiry}

\item{idToken}{The GITKit token of the authenticated user}

\item{localId}{The list of local ID's of the users to inquiry}
}
\value{
IdentitytoolkitRelyingpartyGetAccountInfoRequest object
}
\description{
IdentitytoolkitRelyingpartyGetAccountInfoRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to get the account information.
}
\seealso{
Other IdentitytoolkitRelyingpartyGetAccountInfoRequest functions: \code{\link{relyingparty.getAccountInfo}}
}

