% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyGetPublicKeysResponse}
\alias{IdentitytoolkitRelyingpartyGetPublicKeysResponse}
\title{IdentitytoolkitRelyingpartyGetPublicKeysResponse Object}
\usage{
IdentitytoolkitRelyingpartyGetPublicKeysResponse()
}
\value{
IdentitytoolkitRelyingpartyGetPublicKeysResponse object
}
\description{
IdentitytoolkitRelyingpartyGetPublicKeysResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Respone of getting public keys.
}

