% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyResetPasswordRequest}
\alias{IdentitytoolkitRelyingpartyResetPasswordRequest}
\title{IdentitytoolkitRelyingpartyResetPasswordRequest Object}
\usage{
IdentitytoolkitRelyingpartyResetPasswordRequest(email = NULL,
  newPassword = NULL, oldPassword = NULL, oobCode = NULL)
}
\arguments{
\item{email}{The email address of the user}

\item{newPassword}{The new password inputted by the user}

\item{oldPassword}{The old password inputted by the user}

\item{oobCode}{The confirmation code}
}
\value{
IdentitytoolkitRelyingpartyResetPasswordRequest object
}
\description{
IdentitytoolkitRelyingpartyResetPasswordRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to reset the password.
}
\seealso{
Other IdentitytoolkitRelyingpartyResetPasswordRequest functions: \code{\link{relyingparty.resetPassword}}
}
