% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartySetAccountInfoRequest}
\alias{IdentitytoolkitRelyingpartySetAccountInfoRequest}
\title{IdentitytoolkitRelyingpartySetAccountInfoRequest Object}
\usage{
IdentitytoolkitRelyingpartySetAccountInfoRequest(captchaChallenge = NULL,
  captchaResponse = NULL, createdAt = NULL, delegatedProjectNumber = NULL,
  deleteAttribute = NULL, deleteProvider = NULL, disableUser = NULL,
  displayName = NULL, email = NULL, emailVerified = NULL,
  idToken = NULL, instanceId = NULL, lastLoginAt = NULL, localId = NULL,
  oobCode = NULL, password = NULL, photoUrl = NULL, provider = NULL,
  returnSecureToken = NULL, upgradeToFederatedLogin = NULL,
  validSince = NULL)
}
\arguments{
\item{captchaChallenge}{The captcha challenge}

\item{captchaResponse}{Response to the captcha}

\item{createdAt}{The timestamp when the account is created}

\item{delegatedProjectNumber}{GCP project number of the requesting delegated app}

\item{deleteAttribute}{The attributes users request to delete}

\item{deleteProvider}{The IDPs the user request to delete}

\item{disableUser}{Whether to disable the user}

\item{displayName}{The name of the user}

\item{email}{The email of the user}

\item{emailVerified}{Mark the email as verified or not}

\item{idToken}{The GITKit token of the authenticated user}

\item{instanceId}{Instance id token of the app}

\item{lastLoginAt}{Last login timestamp}

\item{localId}{The local ID of the user}

\item{oobCode}{The out-of-band code of the change email request}

\item{password}{The new password of the user}

\item{photoUrl}{The photo url of the user}

\item{provider}{The associated IDPs of the user}

\item{returnSecureToken}{Whether return sts id token and refresh token instead of gitkit token}

\item{upgradeToFederatedLogin}{Mark the user to upgrade to federated login}

\item{validSince}{Timestamp in seconds for valid login token}
}
\value{
IdentitytoolkitRelyingpartySetAccountInfoRequest object
}
\description{
IdentitytoolkitRelyingpartySetAccountInfoRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to set the account information.
}
\seealso{
Other IdentitytoolkitRelyingpartySetAccountInfoRequest functions: \code{\link{relyingparty.setAccountInfo}}
}
