% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartySetProjectConfigRequest}
\alias{IdentitytoolkitRelyingpartySetProjectConfigRequest}
\title{IdentitytoolkitRelyingpartySetProjectConfigRequest Object}
\usage{
IdentitytoolkitRelyingpartySetProjectConfigRequest(allowPasswordUser = NULL,
  apiKey = NULL, authorizedDomains = NULL, changeEmailTemplate = NULL,
  delegatedProjectNumber = NULL, enableAnonymousUser = NULL,
  idpConfig = NULL, legacyResetPasswordTemplate = NULL,
  resetPasswordTemplate = NULL, useEmailSending = NULL,
  verifyEmailTemplate = NULL)
}
\arguments{
\item{allowPasswordUser}{Whether to allow password user sign in or sign up}

\item{apiKey}{Browser API key, needed when making http request to Apiary}

\item{authorizedDomains}{Authorized domains for widget redirect}

\item{changeEmailTemplate}{Change email template}

\item{delegatedProjectNumber}{GCP project number of the requesting delegated app}

\item{enableAnonymousUser}{Whether to enable anonymous user}

\item{idpConfig}{Oauth2 provider configuration}

\item{legacyResetPasswordTemplate}{Legacy reset password email template}

\item{resetPasswordTemplate}{Reset password email template}

\item{useEmailSending}{Whether to use email sending provided by Firebear}

\item{verifyEmailTemplate}{Verify email template}
}
\value{
IdentitytoolkitRelyingpartySetProjectConfigRequest object
}
\description{
IdentitytoolkitRelyingpartySetProjectConfigRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to set the project configuration.
}
\seealso{
Other IdentitytoolkitRelyingpartySetProjectConfigRequest functions: \code{\link{relyingparty.setProjectConfig}}
}

