% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartySetProjectConfigResponse}
\alias{IdentitytoolkitRelyingpartySetProjectConfigResponse}
\title{IdentitytoolkitRelyingpartySetProjectConfigResponse Object}
\usage{
IdentitytoolkitRelyingpartySetProjectConfigResponse(projectId = NULL)
}
\arguments{
\item{projectId}{Project ID of the relying party}
}
\value{
IdentitytoolkitRelyingpartySetProjectConfigResponse object
}
\description{
IdentitytoolkitRelyingpartySetProjectConfigResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of setting the project configuration.
}

