% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartySignOutUserRequest}
\alias{IdentitytoolkitRelyingpartySignOutUserRequest}
\title{IdentitytoolkitRelyingpartySignOutUserRequest Object}
\usage{
IdentitytoolkitRelyingpartySignOutUserRequest(instanceId = NULL,
  localId = NULL)
}
\arguments{
\item{instanceId}{Instance id token of the app}

\item{localId}{The local ID of the user}
}
\value{
IdentitytoolkitRelyingpartySignOutUserRequest object
}
\description{
IdentitytoolkitRelyingpartySignOutUserRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to sign out user.
}
\seealso{
Other IdentitytoolkitRelyingpartySignOutUserRequest functions: \code{\link{relyingparty.signOutUser}}
}

