% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartySignOutUserResponse}
\alias{IdentitytoolkitRelyingpartySignOutUserResponse}
\title{IdentitytoolkitRelyingpartySignOutUserResponse Object}
\usage{
IdentitytoolkitRelyingpartySignOutUserResponse(localId = NULL)
}
\arguments{
\item{localId}{The local ID of the user}
}
\value{
IdentitytoolkitRelyingpartySignOutUserResponse object
}
\description{
IdentitytoolkitRelyingpartySignOutUserResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of signing out user.
}
