% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyVerifyAssertionRequest}
\alias{IdentitytoolkitRelyingpartyVerifyAssertionRequest}
\title{IdentitytoolkitRelyingpartyVerifyAssertionRequest Object}
\usage{

  IdentitytoolkitRelyingpartyVerifyAssertionRequest(delegatedProjectNumber = NULL,
  idToken = NULL, instanceId = NULL, pendingIdToken = NULL,
  postBody = NULL, requestUri = NULL, returnIdpCredential = NULL,
  returnRefreshToken = NULL, returnSecureToken = NULL, sessionId = NULL)
}
\arguments{
\item{delegatedProjectNumber}{GCP project number of the requesting delegated app}

\item{idToken}{The GITKit token of the authenticated user}

\item{instanceId}{Instance id token of the app}

\item{pendingIdToken}{The GITKit token for the non-trusted IDP pending to be confirmed by the user}

\item{postBody}{The post body if the request is a HTTP POST}

\item{requestUri}{The URI to which the IDP redirects the user back}

\item{returnIdpCredential}{Whether return 200 and IDP credential rather than throw exception when federated id is already linked}

\item{returnRefreshToken}{Whether to return refresh tokens}

\item{returnSecureToken}{Whether return sts id token and refresh token instead of gitkit token}

\item{sessionId}{Session ID, which should match the one in previous createAuthUri request}
}
\value{
IdentitytoolkitRelyingpartyVerifyAssertionRequest object
}
\description{
IdentitytoolkitRelyingpartyVerifyAssertionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to verify the IDP assertion.
}
\seealso{
Other IdentitytoolkitRelyingpartyVerifyAssertionRequest functions: \code{\link{relyingparty.verifyAssertion}}
}
