% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyVerifyCustomTokenRequest}
\alias{IdentitytoolkitRelyingpartyVerifyCustomTokenRequest}
\title{IdentitytoolkitRelyingpartyVerifyCustomTokenRequest Object}
\usage{

  IdentitytoolkitRelyingpartyVerifyCustomTokenRequest(delegatedProjectNumber = NULL,
  instanceId = NULL, returnSecureToken = NULL, token = NULL)
}
\arguments{
\item{delegatedProjectNumber}{GCP project number of the requesting delegated app}

\item{instanceId}{Instance id token of the app}

\item{returnSecureToken}{Whether return sts id token and refresh token instead of gitkit token}

\item{token}{The custom token to verify}
}
\value{
IdentitytoolkitRelyingpartyVerifyCustomTokenRequest object
}
\description{
IdentitytoolkitRelyingpartyVerifyCustomTokenRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to verify a custom token
}
\seealso{
Other IdentitytoolkitRelyingpartyVerifyCustomTokenRequest functions: \code{\link{relyingparty.verifyCustomToken}}
}
