% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdentitytoolkitRelyingpartyVerifyPasswordRequest}
\alias{IdentitytoolkitRelyingpartyVerifyPasswordRequest}
\title{IdentitytoolkitRelyingpartyVerifyPasswordRequest Object}
\usage{
IdentitytoolkitRelyingpartyVerifyPasswordRequest(captchaChallenge = NULL,
  captchaResponse = NULL, delegatedProjectNumber = NULL, email = NULL,
  idToken = NULL, instanceId = NULL, password = NULL,
  pendingIdToken = NULL, returnSecureToken = NULL)
}
\arguments{
\item{captchaChallenge}{The captcha challenge}

\item{captchaResponse}{Response to the captcha}

\item{delegatedProjectNumber}{GCP project number of the requesting delegated app}

\item{email}{The email of the user}

\item{idToken}{The GITKit token of the authenticated user}

\item{instanceId}{Instance id token of the app}

\item{password}{The password inputed by the user}

\item{pendingIdToken}{The GITKit token for the non-trusted IDP, which is to be confirmed by the user}

\item{returnSecureToken}{Whether return sts id token and refresh token instead of gitkit token}
}
\value{
IdentitytoolkitRelyingpartyVerifyPasswordRequest object
}
\description{
IdentitytoolkitRelyingpartyVerifyPasswordRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to verify the password.
}
\seealso{
Other IdentitytoolkitRelyingpartyVerifyPasswordRequest functions: \code{\link{relyingparty.verifyPassword}}
}
