% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{IdpConfig}
\alias{IdpConfig}
\title{IdpConfig Object}
\usage{
IdpConfig(clientId = NULL, enabled = NULL, experimentPercent = NULL,
  provider = NULL, secret = NULL, whitelistedAudiences = NULL)
}
\arguments{
\item{clientId}{OAuth2 client ID}

\item{enabled}{Whether this IDP is enabled}

\item{experimentPercent}{Percent of users who will be prompted/redirected federated login for this IDP}

\item{provider}{OAuth2 provider}

\item{secret}{OAuth2 client secret}

\item{whitelistedAudiences}{Whitelisted client IDs for audience check}
}
\value{
IdpConfig object
}
\description{
IdpConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for a single idp configuration.
}

