% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{relyingparty.getOobConfirmationCode}
\alias{relyingparty.getOobConfirmationCode}
\title{Get a code for user action confirmation.}
\usage{
relyingparty.getOobConfirmationCode(Relyingparty)
}
\arguments{
\item{Relyingparty}{The \link{Relyingparty} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/identity-toolkit/v3/}{Google Documentation}

Other Relyingparty functions: \code{\link{Relyingparty}}
}

