% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{relyingparty.getProjectConfig}
\alias{relyingparty.getProjectConfig}
\title{Get project configuration.}
\usage{
relyingparty.getProjectConfig(delegatedProjectNumber = NULL,
  projectNumber = NULL)
}
\arguments{
\item{delegatedProjectNumber}{Delegated GCP project number of the request}

\item{projectNumber}{GCP project number of the request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/identity-toolkit/v3/}{Google Documentation}
}
