% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{Relyingparty}
\alias{Relyingparty}
\title{Relyingparty Object}
\usage{
Relyingparty(captchaResp = NULL, challenge = NULL, email = NULL,
  idToken = NULL, newEmail = NULL, requestType = NULL, userIp = NULL)
}
\arguments{
\item{captchaResp}{The recaptcha response from the user}

\item{challenge}{The recaptcha challenge presented to the user}

\item{email}{The email of the user}

\item{idToken}{The user's Gitkit login token for email change}

\item{newEmail}{The new email if the code is for email change}

\item{requestType}{The request type}

\item{userIp}{The IP address of the user}
}
\value{
Relyingparty object
}
\description{
Relyingparty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request of getting a code for user confirmation (reset password, change email etc.)
}
\seealso{
Other Relyingparty functions: \code{\link{relyingparty.getOobConfirmationCode}}
}
