% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{relyingparty.resetPassword}
\alias{relyingparty.resetPassword}
\title{Reset password for a user.}
\usage{
relyingparty.resetPassword(IdentitytoolkitRelyingpartyResetPasswordRequest)
}
\arguments{
\item{IdentitytoolkitRelyingpartyResetPasswordRequest}{The \link{IdentitytoolkitRelyingpartyResetPasswordRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/identity-toolkit/v3/}{Google Documentation}

Other IdentitytoolkitRelyingpartyResetPasswordRequest functions: \code{\link{IdentitytoolkitRelyingpartyResetPasswordRequest}}
}

