% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{relyingparty.setProjectConfig}
\alias{relyingparty.setProjectConfig}
\title{Set project configuration.}
\usage{

  relyingparty.setProjectConfig(IdentitytoolkitRelyingpartySetProjectConfigRequest)
}
\arguments{
\item{IdentitytoolkitRelyingpartySetProjectConfigRequest}{The \link{IdentitytoolkitRelyingpartySetProjectConfigRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/identity-toolkit/v3/}{Google Documentation}

Other IdentitytoolkitRelyingpartySetProjectConfigRequest functions: \code{\link{IdentitytoolkitRelyingpartySetProjectConfigRequest}}
}

