% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{relyingparty.verifyAssertion}
\alias{relyingparty.verifyAssertion}
\title{Verifies the assertion returned by the IdP.}
\usage{
relyingparty.verifyAssertion(IdentitytoolkitRelyingpartyVerifyAssertionRequest)
}
\arguments{
\item{IdentitytoolkitRelyingpartyVerifyAssertionRequest}{The \link{IdentitytoolkitRelyingpartyVerifyAssertionRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/identity-toolkit/v3/}{Google Documentation}

Other IdentitytoolkitRelyingpartyVerifyAssertionRequest functions: \code{\link{IdentitytoolkitRelyingpartyVerifyAssertionRequest}}
}
