% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\name{relyingparty.verifyCustomToken}
\alias{relyingparty.verifyCustomToken}
\title{Verifies the developer asserted ID token.}
\usage{

  relyingparty.verifyCustomToken(IdentitytoolkitRelyingpartyVerifyCustomTokenRequest)
}
\arguments{
\item{IdentitytoolkitRelyingpartyVerifyCustomTokenRequest}{The \link{IdentitytoolkitRelyingpartyVerifyCustomTokenRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/identity-toolkit/v3/}{Google Documentation}

Other IdentitytoolkitRelyingpartyVerifyCustomTokenRequest functions: \code{\link{IdentitytoolkitRelyingpartyVerifyCustomTokenRequest}}
}

