% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{SetAccountInfoResponse}
\alias{SetAccountInfoResponse}
\title{SetAccountInfoResponse Object}
\usage{
SetAccountInfoResponse(SetAccountInfoResponse.providerUserInfo = NULL,
  displayName = NULL, email = NULL, expiresIn = NULL, idToken = NULL,
  localId = NULL, newEmail = NULL, passwordHash = NULL, photoUrl = NULL,
  providerUserInfo = NULL, refreshToken = NULL)
}
\arguments{
\item{SetAccountInfoResponse.providerUserInfo}{The \link{SetAccountInfoResponse.providerUserInfo} object or list of objects}

\item{displayName}{The name of the user}

\item{email}{The email of the user}

\item{expiresIn}{If idToken is STS id token, then this field will be expiration time of STS id token in seconds}

\item{idToken}{The Gitkit id token to login the newly sign up user}

\item{localId}{The local ID of the user}

\item{newEmail}{The new email the user attempts to change to}

\item{passwordHash}{The user's hashed password}

\item{photoUrl}{The photo url of the user}

\item{providerUserInfo}{The user's profiles at the associated IdPs}

\item{refreshToken}{If idToken is STS id token, then this field will be refresh token}
}
\value{
SetAccountInfoResponse object
}
\description{
SetAccountInfoResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Respone of setting the account information.
}
\seealso{
Other SetAccountInfoResponse functions: \code{\link{SetAccountInfoResponse.providerUserInfo}}
}

