% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{SignupNewUserResponse}
\alias{SignupNewUserResponse}
\title{SignupNewUserResponse Object}
\usage{
SignupNewUserResponse(displayName = NULL, email = NULL, expiresIn = NULL,
  idToken = NULL, localId = NULL, refreshToken = NULL)
}
\arguments{
\item{displayName}{The name of the user}

\item{email}{The email of the user}

\item{expiresIn}{If idToken is STS id token, then this field will be expiration time of STS id token in seconds}

\item{idToken}{The Gitkit id token to login the newly sign up user}

\item{localId}{The RP local ID of the user}

\item{refreshToken}{If idToken is STS id token, then this field will be refresh token}
}
\value{
SignupNewUserResponse object
}
\description{
SignupNewUserResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of signing up new user, creating anonymous user or anonymous user reauth.
}
