% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{UploadAccountResponse}
\alias{UploadAccountResponse}
\title{UploadAccountResponse Object}
\usage{
UploadAccountResponse(UploadAccountResponse.error = NULL, error = NULL)
}
\arguments{
\item{UploadAccountResponse.error}{The \link{UploadAccountResponse.error} object or list of objects}

\item{error}{The error encountered while processing the account info}
}
\value{
UploadAccountResponse object
}
\description{
UploadAccountResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Respone of uploading accounts in batch.
}
\seealso{
Other UploadAccountResponse functions: \code{\link{UploadAccountResponse.error}}
}
