% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{UserInfo.providerUserInfo}
\alias{UserInfo.providerUserInfo}
\title{UserInfo.providerUserInfo Object}
\usage{
UserInfo.providerUserInfo()
}
\value{
UserInfo.providerUserInfo object
}
\description{
UserInfo.providerUserInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other UserInfo functions: \code{\link{UserInfo}}
}

