% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{UserInfo}
\alias{UserInfo}
\title{UserInfo Object}
\usage{
UserInfo(UserInfo.providerUserInfo = NULL, createdAt = NULL,
  disabled = NULL, displayName = NULL, email = NULL,
  emailVerified = NULL, lastLoginAt = NULL, localId = NULL,
  passwordHash = NULL, passwordUpdatedAt = NULL, photoUrl = NULL,
  providerUserInfo = NULL, salt = NULL, screenName = NULL,
  validSince = NULL, version = NULL)
}
\arguments{
\item{UserInfo.providerUserInfo}{The \link{UserInfo.providerUserInfo} object or list of objects}

\item{createdAt}{User creation timestamp}

\item{disabled}{Whether the user is disabled}

\item{displayName}{The name of the user}

\item{email}{The email of the user}

\item{emailVerified}{Whether the email has been verified}

\item{lastLoginAt}{last login timestamp}

\item{localId}{The local ID of the user}

\item{passwordHash}{The user's hashed password}

\item{passwordUpdatedAt}{The timestamp when the password was last updated}

\item{photoUrl}{The URL of the user profile photo}

\item{providerUserInfo}{The IDP of the user}

\item{salt}{The user's password salt}

\item{screenName}{User's screen name at Twitter}

\item{validSince}{Timestamp in seconds for valid login token}

\item{version}{Version of the user's password}
}
\value{
UserInfo object
}
\description{
UserInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for an individual account info.
}
\seealso{
Other UserInfo functions: \code{\link{UserInfo.providerUserInfo}}
}

