% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{VerifyCustomTokenResponse}
\alias{VerifyCustomTokenResponse}
\title{VerifyCustomTokenResponse Object}
\usage{
VerifyCustomTokenResponse(expiresIn = NULL, idToken = NULL,
  refreshToken = NULL)
}
\arguments{
\item{expiresIn}{If idToken is STS id token, then this field will be expiration time of STS id token in seconds}

\item{idToken}{The GITKit token for authenticated user}

\item{refreshToken}{If idToken is STS id token, then this field will be refresh token}
}
\value{
VerifyCustomTokenResponse object
}
\description{
VerifyCustomTokenResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response from verifying a custom token
}
