% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{VerifyPasswordResponse}
\alias{VerifyPasswordResponse}
\title{VerifyPasswordResponse Object}
\usage{
VerifyPasswordResponse(displayName = NULL, email = NULL, expiresIn = NULL,
  idToken = NULL, localId = NULL, oauthAccessToken = NULL,
  oauthAuthorizationCode = NULL, oauthExpireIn = NULL, photoUrl = NULL,
  refreshToken = NULL, registered = NULL)
}
\arguments{
\item{displayName}{The name of the user}

\item{email}{The email returned by the IdP}

\item{expiresIn}{If idToken is STS id token, then this field will be expiration time of STS id token in seconds}

\item{idToken}{The GITKit token for authenticated user}

\item{localId}{The RP local ID if it's already been mapped to the IdP account identified by the federated ID}

\item{oauthAccessToken}{The OAuth2 access token}

\item{oauthAuthorizationCode}{The OAuth2 authorization code}

\item{oauthExpireIn}{The lifetime in seconds of the OAuth2 access token}

\item{photoUrl}{The URI of the user's photo at IdP}

\item{refreshToken}{If idToken is STS id token, then this field will be refresh token}

\item{registered}{Whether the email is registered}
}
\value{
VerifyPasswordResponse object
}
\description{
VerifyPasswordResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request of verifying the password.
}

