% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_functions.R
\docType{package}
\name{identitytoolkit_googleAuthR}
\alias{identitytoolkit_googleAuthR}
\alias{identitytoolkit_googleAuthR-package}
\title{Google Identity Toolkit API
Help the third party sites to implement federated login.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:23:36
filename: /Users/mark/dev/R/autoGoogleAPI/googleidentitytoolkitv3.auto/R/identitytoolkit_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

