% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitytoolkit_objects.R
\name{ResetPasswordResponse}
\alias{ResetPasswordResponse}
\title{ResetPasswordResponse Object}
\usage{
ResetPasswordResponse(email = NULL, newEmail = NULL, requestType = NULL)
}
\arguments{
\item{email}{The user's email}

\item{newEmail}{If the out-of-band code is for email recovery, the user's new email}

\item{requestType}{The request type}
}
\value{
ResetPasswordResponse object
}
\description{
ResetPasswordResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response of resetting the password.
}
